% Auto-generated: do not edit by hand
\name{accordion}

\alias{accordion}

\title{Accordion component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
accordion(children=NULL, id=NULL, label=NULL, labelClass=NULL,
isOpenOnFirstRender=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. }

\item{labelClass}{Character. }

\item{isOpenOnFirstRender}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

