% Auto-generated: do not edit by hand
\name{searchBar}

\alias{searchBar}

\title{SearchBar component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
searchBar(id=NULL, opts=NULL, value=NULL, fireOnSelect=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{opts}{}

\item{value}{Lists containing elements 'value', 'type'.
those elements have the following types:
  - value (character | numeric; optional)
  - type (character; optional). }

\item{fireOnSelect}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

