% Auto-generated: do not edit by hand
\name{treemap}

\alias{treemap}

\title{Treemap component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
treemap(id=NULL, data=NULL, width=NULL, height=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Logical | numeric | character | named list | unnamed list. }

\item{width}{Numeric. }

\item{height}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

